INCOMPLETE: Try replacing the `unsafe-run0!h!` slot with a procedure that doesn't return holes of degree less than the sensitivity degree. This makes some parts of the code much more elegant, especially the way it gives us easy tail calls in `bind!h`, but... it looks like we'll have to push a special hyperstack frame at the start of calls to `run1!h!` and pop it again at the end, which seems... Well, it relies on non-monotone degree-0 effects again, which makes me think this whole system is practically degree-0-sensitive, and I get to wondering what makes this system more principled than mutable variables anyway. This leads me to several thoughts:
Degree-2-sensitivity and degree-3-sensitivity *would* be distinguishable if we had operations that were syntactically of degree 2 or 3, because then we would care about the evaluation order of their higher-degree holes.

Is the real reason I'm asking the programmer to specify a finite usage degree for each effect ahead of time... only that I want to understand hypersnippets better through this exercise?

When you think about it, imperative effects like mutation are actually degree-1 (not degree-0 as we currently say they are in Effection) because you call to them and return again. If you called an effect and it never returned -- like `goto` -- that would be a real degree-0 effect. Perhaps run time errors and nontermination are related to degree-0 effects.

I really oughta put these thoughts in a comment at least, not a commit message.

These changes were written the night of March 7 but left uncommitted for a while.